+++ spec/dummy/client/entries/app-kitten.js
@@ -187,6 +187,7 @@ import { KarlMenu,
          KarlMenuContent } from 'kitten/karl/menus/karl-menu'
 import { KarlBrowsing } from 'app/browsing/browsing'
 import { KarlSeparator } from 'kitten/karl/layout/separator'
+import { KarlImageCropper } from 'kitten/karl/images/image-cropper'
 
 ReactOnRails.register({
   // Box
@@ -370,4 +371,5 @@ ReactOnRails.register({
   KarlBrowsing,
   KarlGithubRibbon,
   KarlSeparator,
+  KarlImageCropper,
 })